Generalize over scroll target.
- Bug report https://bugs.chromium.org/p/chromium/issues/detail?id=157855

- Description of fix from
https://developer.mozilla.org/en-US/docs/Web/API/document/scrollingElement...

"The scrollingElement read-only property of the Document interface returns a reference to the Element that scrolls the document. In standards mode, this is the root element of the document, document.documentElement.

When in quirks mode, the scrollingElement attribute returns the HTML
body element (or null if it does not exist)."